Created a new processors package
This reorganization should keep the files smaller and a bit more maintainable.

XCOM-338
