+++ ecommerce/settings/_oscar.py
@@ -94,8 +94,8 @@ OSCAR_DEFAULT_CURRENCY = 'USD'
 
 # PAYMENT PROCESSING
 PAYMENT_PROCESSORS = (
-    'ecommerce.extensions.payment.processors.Cybersource',
-    'ecommerce.extensions.payment.processors.Paypal',
+    'ecommerce.extensions.payment.processors.cybersource.Cybersource',
+    'ecommerce.extensions.payment.processors.paypal.Paypal',
 )
 
 PAYMENT_PROCESSOR_CONFIG = {
